ai-service/requirements.txt
fastapi
uvicorn
pydantic
openai
google-generativeai>=0.5.0
python-multipart
requests
